Move the Encrypter from the config to the keystore
