+++ crates/handlers/src/views/account/emails/mod.rs
@@ -22,9 +22,9 @@ use mas_axum_utils::{
     csrf::{CsrfExt, ProtectedForm},
     FancyError, SessionInfoExt,
 };
-use mas_config::Encrypter;
 use mas_data_model::{BrowserSession, User, UserEmail};
 use mas_email::Mailer;
+use mas_keystore::Encrypter;
 use mas_router::Route;
 use mas_storage::{
     user::{
